feat: migrate from SQLite to DuckDB
- Replace sqlite:/// with duckdb:/// engine URIs
- Add duckdb-engine dependency to pyproject.toml
- Change Column(Integer) to Column(BigInteger) for all PKs/FKs
- Change Column(Float) to Column(Double) for all numeric columns
- Add Sequence() for auto-increment primary keys (DuckDB has no SERIAL)
- Change external_id from BigInteger to String on orders
- Add TimeZone=UTC config to DuckDB engine connection
- Add timezone normalization event listener for loaded models
- Fix sqlalchemy.orm.mapper deprecation to Mapper
- Add global conftest.py with DuckDB cleanup fixture (gc.collect)
- Improve tearDown: close_all_sessions + dispose + gc.collect
- Update test assertions for .duckdb file extensions
- Update AWS state_handler for .duckdb filename

+++ investing_algorithm_framework/infrastructure/models/order_trade_association.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Table, Column, Integer, ForeignKey
+from sqlalchemy import Table, Column, BigInteger, ForeignKey
 from investing_algorithm_framework.infrastructure.database import SQLBaseModel
 
 # Association table
 order_trade_association = Table(
     'order_trade',  # Table name
     SQLBaseModel.metadata,
-    Column('order_id', Integer, ForeignKey('orders.id'), primary_key=True),
-    Column('trade_id', Integer, ForeignKey('trades.id'), primary_key=True)
+    Column('order_id', BigInteger, ForeignKey('orders.id'), primary_key=True),
+    Column('trade_id', BigInteger, ForeignKey('trades.id'), primary_key=True)
 )
